remove hardcoded retry stuff; put -Werror back in (who the fk removed that?)
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky1@1234 7f574dfc-610e-0410-a909-a81674777703

+++ src/conky.h
@@ -156,7 +156,7 @@ struct mail_s {			// for imap and pop3
 	unsigned long used;
 	unsigned long quota;
 	unsigned long port;
-	unsigned long retries;
+	unsigned int retries;
 	float interval;
 	double last_update;
 	char host[128];
